requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
face-recognition
python-jose
passlib[bcrypt]
pydantic
